+++ tests/typed.py
@@ -18,7 +18,6 @@ from typing import (
     MutableSequence,
     MutableSet,
     NewType,
-    Optional,
     Sequence,
     Set,
     Tuple,
@@ -49,6 +48,7 @@ from hypothesis.strategies import (
     tuples,
 )
 
+from . import FeatureFlag
 from .untyped import gen_attr_names, make_class
 
 PosArg = Any
@@ -58,10 +58,10 @@ T = TypeVar("T")
 
 
 def simple_typed_classes(
-    defaults=None,
+    defaults: FeatureFlag = "sometimes",
     min_attrs=0,
     frozen=False,
-    kw_only=None,
+    kw_only: FeatureFlag = "sometimes",
     newtypes=True,
     text_codec: str = "utf8",
     allow_infinity=None,
@@ -81,7 +81,11 @@ def simple_typed_classes(
 
 
 def simple_typed_dataclasses(
-    defaults=None, min_attrs=0, frozen=False, newtypes=True, allow_nan=True
+    defaults: FeatureFlag = "sometimes",
+    min_attrs=0,
+    frozen=False,
+    newtypes=True,
+    allow_nan=True,
 ):
     """Yield tuples of (class, values)."""
     return lists_of_typed_attrs(
@@ -95,7 +99,11 @@ def simple_typed_dataclasses(
 
 
 def simple_typed_classes_and_strats(
-    defaults=None, min_attrs=0, kw_only=None, newtypes=True, allow_nan=True
+    defaults: FeatureFlag = "sometimes",
+    min_attrs=0,
+    kw_only: FeatureFlag = "sometimes",
+    newtypes=True,
+    allow_nan=True,
 ) -> SearchStrategy[tuple[type, SearchStrategy[PosArgs], SearchStrategy[KwArgs]]]:
     """Yield tuples of (class, (strategies))."""
     return lists_of_typed_attrs(
@@ -108,11 +116,11 @@ def simple_typed_classes_and_strats(
 
 
 def lists_of_typed_attrs(
-    defaults=None,
+    defaults: FeatureFlag = "sometimes",
     min_size=0,
     for_frozen=False,
     allow_mutable_defaults=True,
-    kw_only=None,
+    kw_only: FeatureFlag = "sometimes",
     newtypes=True,
     text_codec="utf8",
     allow_infinity=None,
@@ -140,10 +148,10 @@ def lists_of_typed_attrs(
 
 
 def simple_typed_attrs(
-    defaults=None,
+    defaults: FeatureFlag = "sometimes",
     for_frozen=False,
     allow_mutable_defaults=True,
-    kw_only=None,
+    kw_only: FeatureFlag = "sometimes",
     newtypes=True,
     text_codec="utf8",
     allow_infinity=None,
@@ -301,55 +309,70 @@ def _create_hyp_class_and_strat(
 
 @composite
 def any_typed_attrs(
-    draw: DrawFn, defaults=None, kw_only=None
+    draw: DrawFn,
+    defaults: FeatureFlag = "sometimes",
+    kw_only: FeatureFlag = "sometimes",
 ) -> tuple[_CountingAttr, SearchStrategy[None]]:
     """Attributes typed as `Any`, having values of `None`."""
     default = NOTHING
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default = None
     return (
         field(
             type=Any,
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         just(None),
     )
 
 
 @composite
-def int_typed_attrs(draw, defaults=None, kw_only=None):
+def int_typed_attrs(
+    draw, defaults: FeatureFlag = "sometimes", kw_only: FeatureFlag = "sometimes"
+):
     """
     Generate a tuple of an attribute and a strategy that yields ints for that
     attribute.
     """
     default = NOTHING
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default = draw(integers())
     return (
         field(
             type=int,
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         integers(),
     )
 
 
 @composite
-def str_typed_attrs(draw, defaults=None, kw_only=None, codec: str = "utf8"):
+def str_typed_attrs(
+    draw,
+    defaults: FeatureFlag = "sometimes",
+    kw_only: FeatureFlag = "sometimes",
+    codec: str = "utf8",
+):
     """
     Generate a tuple of an attribute and a strategy that yields strs for that
     attribute.
     """
     default = NOTHING
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default = draw(text())
     return (
         field(
             type=str,
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         text(characters(codec=codec)),
     )
@@ -357,20 +380,26 @@ def str_typed_attrs(draw, defaults=None, kw_only=None, codec: str = "utf8"):
 
 @composite
 def float_typed_attrs(
-    draw, defaults=None, kw_only=None, allow_infinity=None, allow_nan=True
+    draw,
+    defaults: FeatureFlag = "sometimes",
+    kw_only: FeatureFlag = "sometimes",
+    allow_infinity=None,
+    allow_nan=True,
 ):
     """
     Generate a tuple of an attribute and a strategy that yields floats for that
     attribute.
     """
     default = NOTHING
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default = draw(floats(allow_infinity=allow_infinity, allow_nan=allow_nan))
     return (
         field(
             type=float,
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         floats(allow_infinity=allow_infinity, allow_nan=allow_nan),
     )
@@ -378,7 +407,9 @@ def float_typed_attrs(
 
 @composite
 def path_typed_attrs(
-    draw: DrawFn, defaults: Optional[bool] = None, kw_only: Optional[bool] = None
+    draw: DrawFn,
+    defaults: FeatureFlag = "sometimes",
+    kw_only: FeatureFlag = "sometimes",
 ) -> tuple[_CountingAttr, SearchStrategy[Path]]:
     """
     Generate a tuple of an attribute and a strategy that yields paths for that
@@ -386,13 +417,15 @@ def path_typed_attrs(
     """
 
     default = NOTHING
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default = Path(draw(text(ascii_lowercase, min_size=1)))
     return (
         field(
             type=Path,
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         text(ascii_lowercase, min_size=1).map(Path),
     )
@@ -400,7 +433,10 @@ def path_typed_attrs(
 
 @composite
 def dict_typed_attrs(
-    draw: DrawFn, defaults=None, allow_mutable_defaults=True, kw_only=None
+    draw: DrawFn,
+    defaults: FeatureFlag = "sometimes",
+    allow_mutable_defaults=True,
+    kw_only: FeatureFlag = "sometimes",
 ) -> tuple[_CountingAttr, SearchStrategy[dict[str, int]]]:
     """
     Generate a tuple of an attribute and a strategy that yields dictionaries
@@ -409,7 +445,7 @@ def dict_typed_attrs(
     """
     default = NOTHING
     val_strat = dictionaries(keys=text(), values=integers())
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default_val = draw(val_strat)
         if not allow_mutable_defaults or draw(booleans()):
             default = Factory(lambda: default_val)
@@ -420,7 +456,9 @@ def dict_typed_attrs(
         field(
             type=type,
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         val_strat,
     )
@@ -428,7 +466,10 @@ def dict_typed_attrs(
 
 @composite
 def new_dict_typed_attrs(
-    draw, defaults=None, allow_mutable_defaults=True, kw_only=None
+    draw,
+    defaults: FeatureFlag = "sometimes",
+    allow_mutable_defaults=True,
+    kw_only: FeatureFlag = "sometimes",
 ):
     """
     Generate a tuple of an attribute and a strategy that yields dictionaries
@@ -438,7 +479,7 @@ def new_dict_typed_attrs(
     """
     default_val = NOTHING
     val_strat = dictionaries(keys=text(), values=integers())
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default_val = draw(val_strat)
         if not allow_mutable_defaults or draw(booleans()):
             default = Factory(lambda: default_val)
@@ -451,7 +492,9 @@ def new_dict_typed_attrs(
         field(
             type=dict[str, int],
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         val_strat,
     )
@@ -460,10 +503,10 @@ def new_dict_typed_attrs(
 @composite
 def set_typed_attrs(
     draw: DrawFn,
-    defaults=None,
+    defaults: FeatureFlag = "sometimes",
     allow_mutable_defaults=True,
     legacy_types_only=False,
-    kw_only=None,
+    kw_only: FeatureFlag = "sometimes",
 ):
     """
     Generate a tuple of an attribute and a strategy that yields sets
@@ -471,7 +514,7 @@ def set_typed_attrs(
     """
     default_val = NOTHING
     val_strat = sets(integers())
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default_val = draw(val_strat)
         if not allow_mutable_defaults or draw(booleans()):
             default = Factory(lambda: default_val)
@@ -491,7 +534,9 @@ def set_typed_attrs(
         field(
             type=type,
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         val_strat,
     )
@@ -499,7 +544,10 @@ def set_typed_attrs(
 
 @composite
 def frozenset_typed_attrs(
-    draw: DrawFn, defaults=None, legacy_types_only=False, kw_only=None
+    draw: DrawFn,
+    defaults: FeatureFlag = "sometimes",
+    legacy_types_only=False,
+    kw_only: FeatureFlag = "sometimes",
 ):
     """
     Generate a tuple of an attribute and a strategy that yields frozensets
@@ -507,7 +555,7 @@ def frozenset_typed_attrs(
     """
     default = NOTHING
     val_strat = frozensets(integers())
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default = draw(val_strat)
     type = draw(
         sampled_from(
@@ -520,7 +568,9 @@ def frozenset_typed_attrs(
         field(
             type=type,
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         val_strat,
     )
@@ -529,10 +579,10 @@ def frozenset_typed_attrs(
 @composite
 def list_typed_attrs(
     draw: DrawFn,
-    defaults=None,
+    defaults: FeatureFlag = "sometimes",
     allow_mutable_defaults=True,
     legacy_types_only=False,
-    kw_only=None,
+    kw_only: FeatureFlag = "sometimes",
 ) -> tuple[_CountingAttr, SearchStrategy[list[float]]]:
     """
     Generate a tuple of an attribute and a strategy that yields lists
@@ -540,7 +590,7 @@ def list_typed_attrs(
     """
     default_val = NOTHING
     val_strat = lists(floats(allow_infinity=False, allow_nan=False))
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default_val = draw(val_strat)
         if not allow_mutable_defaults or draw(booleans()):
             default = Factory(lambda: default_val)
@@ -558,7 +608,9 @@ def list_typed_attrs(
                 )
             ),
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         val_strat,
     )
@@ -567,10 +619,10 @@ def list_typed_attrs(
 @composite
 def seq_typed_attrs(
     draw,
-    defaults=None,
+    defaults: FeatureFlag = "sometimes",
     allow_mutable_defaults=True,
     legacy_types_only=False,
-    kw_only=None,
+    kw_only: FeatureFlag = "sometimes",
 ):
     """
     Generate a tuple of an attribute and a strategy that yields tuples
@@ -581,7 +633,7 @@ def seq_typed_attrs(
     """
     default_val = NOTHING
     val_strat = lists(integers()).map(tuple)
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default_val = draw(val_strat)
         if not allow_mutable_defaults or draw(booleans()):
             default = Factory(lambda: default_val)
@@ -594,7 +646,9 @@ def seq_typed_attrs(
         field(
             type=AbcSequence[int] if not legacy_types_only else Sequence[int],
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         val_strat,
     )
@@ -603,10 +657,10 @@ def seq_typed_attrs(
 @composite
 def mutable_seq_typed_attrs(
     draw,
-    defaults=None,
+    defaults: FeatureFlag = "sometimes",
     allow_mutable_defaults=True,
     legacy_types_only=False,
-    kw_only=None,
+    kw_only: FeatureFlag = "sometimes",
 ):
     """
     Generate a tuple of an attribute and a strategy that yields lists
@@ -614,7 +668,7 @@ def mutable_seq_typed_attrs(
     """
     default_val = NOTHING
     val_strat = lists(floats(allow_infinity=False, allow_nan=False))
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default_val = draw(val_strat)
         if not allow_mutable_defaults or draw(booleans()):
             default = Factory(lambda: default_val)
@@ -631,21 +685,28 @@ def mutable_seq_typed_attrs(
                 else MutableSequence[float]
             ),
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         val_strat,
     )
 
 
 @composite
-def homo_tuple_typed_attrs(draw, defaults=None, legacy_types_only=False, kw_only=None):
+def homo_tuple_typed_attrs(
+    draw,
+    defaults: FeatureFlag = "sometimes",
+    legacy_types_only=False,
+    kw_only: FeatureFlag = "sometimes",
+):
     """
     Generate a tuple of an attribute and a strategy that yields homogenous
     tuples for that attribute. The tuples contain strings.
     """
     default = NOTHING
     val_strat = tuples(text(), text(), text())
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default = draw(val_strat)
     return (
         field(
@@ -657,34 +718,46 @@ def homo_tuple_typed_attrs(draw, defaults=None, legacy_types_only=False, kw_only
                 )
             ),
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         val_strat,
     )
 
 
 @composite
-def newtype_int_typed_attrs(draw: DrawFn, defaults=None, kw_only=None):
+def newtype_int_typed_attrs(
+    draw: DrawFn,
+    defaults: FeatureFlag = "sometimes",
+    kw_only: FeatureFlag = "sometimes",
+):
     """
     Generate a tuple of an attribute and a strategy that yields ints for that
     attribute.
     """
     default = NOTHING
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default = draw(integers())
     NewInt = NewType("NewInt", int)
     return (
         field(
             type=NewInt,
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         integers(),
     )
 
 
 @composite
-def newtype_attrs_typed_attrs(draw: DrawFn, defaults=None, kw_only=None):
+def newtype_attrs_typed_attrs(
+    draw: DrawFn,
+    defaults: FeatureFlag = "sometimes",
+    kw_only: FeatureFlag = "sometimes",
+):
     """
     Generate a tuple of an attribute and a strategy that yields values for that
     attribute.
@@ -695,7 +768,7 @@ def newtype_attrs_typed_attrs(draw: DrawFn, defaults=None, kw_only=None):
     class NewTypeAttrs:
         a: int
 
-    if defaults is True or (defaults is None and draw(booleans())):
+    if defaults == "always" or (defaults == "sometimes" and draw(booleans())):
         default = NewTypeAttrs(draw(integers()))
 
     NewAttrs = NewType("NewAttrs", NewTypeAttrs)
@@ -703,7 +776,9 @@ def newtype_attrs_typed_attrs(draw: DrawFn, defaults=None, kw_only=None):
         field(
             type=NewAttrs,
             default=default,
-            kw_only=draw(booleans()) if kw_only is None else kw_only,
+            kw_only=(
+                draw(booleans()) if kw_only == "sometimes" else (kw_only == "always")
+            ),
         ),
         integers().map(NewTypeAttrs),
     )
@@ -791,7 +866,10 @@ def dict_of_class(
 
 
 def _create_hyp_nested_strategy(
-    simple_class_strategy: SearchStrategy, kw_only=None, newtypes=True, allow_nan=True
+    simple_class_strategy: SearchStrategy,
+    kw_only: FeatureFlag = "sometimes",
+    newtypes=True,
+    allow_nan=True,
 ) -> SearchStrategy[tuple[type, SearchStrategy[PosArgs], SearchStrategy[KwArgs]]]:
     """
     Create a recursive attrs class.
@@ -840,7 +918,11 @@ def nested_classes(
 
 
 def nested_typed_classes_and_strat(
-    defaults=None, min_attrs=0, kw_only=None, newtypes=True, allow_nan=True
+    defaults: FeatureFlag = "sometimes",
+    min_attrs=0,
+    kw_only: FeatureFlag = "sometimes",
+    newtypes=True,
+    allow_nan=True,
 ) -> SearchStrategy[tuple[type, SearchStrategy[PosArgs]]]:
     return recursive(
         simple_typed_classes_and_strats(
@@ -863,9 +945,9 @@ def nested_typed_classes_and_strat(
 @composite
 def nested_typed_classes(
     draw: DrawFn,
-    defaults=None,
+    defaults: FeatureFlag = "sometimes",
     min_attrs=0,
-    kw_only=None,
+    kw_only: FeatureFlag = "sometimes",
     newtypes=True,
     allow_nan=True,
 ):